+++ lib/tailwind_ui.dart
@@ -3,5 +3,6 @@ library tailwind_ui;
 export 'src/application_ui/elements/avatars/circular_avatar.dart';
 export 'src/application_ui/elements/avatars/circular_avatar_with_notification.dart';
 export 'src/application_ui/elements/avatars/circular_avatar_with_placeholder_icon.dart';
+export 'src/application_ui/elements/avatars/circular_avatar_with_placeholder_initials.dart';
 export 'src/application_ui/elements/avatars/rounded_avatar.dart';
 export 'src/application_ui/elements/avatars/rounded_avatar_with_notification.dart';